first commit - just algorithm @bradduthie
How do I access 'manage_target' from manager.c without it being present in paras vector?
Is my way of accessing the estimated population correct in manager.c?
Can I add a column (#18) to the AGENT array (in initialise.R) to keep track of whether the policy was updated of not?
If yes, I still need to find a way to modify it from manager.c