changing the measure of variance around target
Hi @bradduthie, here are the sim scripts with modified calculus for the cumulated deviation from target: it sums the absolute value of the difference between the actual population size and manager target over all the time steps, and also saves the number of time steps before extinction or end of simulation.
Could you please run the NORESET script for all BB values with the UT values of 0.1 and 0.2 on the one hand; and the MEMORY_NORESET one for {UT=0; BB=0} and all the BB values for the UT of 0.3 on the other hand please?
Cheers!